+++ werkzeug/wrappers.py
@@ -53,6 +53,19 @@ def _run_wsgi_app(*args):
     return _run_wsgi_app(*args)
 
 
+def _warn_if_string(iterable):
+    """Helper for the response objects to check if the iterable returned
+    to the WSGI server is not a string.
+    """
+    if isinstance(iterable, basestring):
+        from warnings import Warning
+        warn(Warning('response iterable was set to a string.  This appears '
+                     'to work but means that the server will send the '
+                     'data to the client char, by char.  This is almost '
+                     'never intended behavior, use response.data to assign '
+                     'strings to the response object.'), stacklevel=2)
+
+
 class BaseRequest(object):
     """Very basic request object.  This does not implement advanced stuff like
     entity tag parsing or cache controls.  The request object is created with
@@ -540,9 +553,8 @@ class BaseResponse(object):
     default_mimetype = 'text/plain'
 
     #: if set to `False` accessing properties on the response object will
-    #: not try to consume the response iterator and buffer it.  By default
-    #: this will automatically happen for properties such as `data`.
-    implicit_iterable_buffering = True
+    #: not try to consume the response iterator and convert it into a list.
+    implicit_seqence_conversion = True
 
     def __init__(self, response=None, status=None, headers=None,
                  mimetype=None, content_type=None, direct_passthrough=False):
@@ -568,16 +580,17 @@ class BaseResponse(object):
         else:
             self.status = status
 
+        self.direct_passthrough = direct_passthrough
+        self._on_close = []
+
         # we set the response after the headers so that if a class changes
         # the charset attribute, the data is set in the correct charset.
         if response is None:
-            self.iterable = []
+            self.response = []
         elif isinstance(response, basestring):
             self.data = response
         else:
-            self.iterable = response
-        self.direct_passthrough = direct_passthrough
-        self._on_close = []
+            self.response = response
 
     def call_on_close(self, func):
         """Adds a function to the internal list of functions that should
@@ -587,20 +600,8 @@ class BaseResponse(object):
         """
         self._on_close.append(func)
 
-    def _get_response(self):
-        from warnings import warn
-        warn(DeprecationWarning('response.respons was renamed to '
-                                'response.iterable'), stacklevel=2)
-        return self.iterable
-    def _set_response(self, value):
-        from warnings import warn
-        warn(DeprecationWarning('response.respons was renamed to '
-                                'response.iterable'), stacklevel=2)
-        self.iterable = value
-    response = property(_get_response, _set_response)
-
     def __repr__(self):
-        if self.iterable_is_buffered:
+        if self.is_sequence:
             body_info = '%d bytes' % sum(map(len, self.iter_encoded()))
         else:
             body_info = self.is_streamed and 'streamed' or 'likely-streamed'
@@ -683,61 +684,74 @@ class BaseResponse(object):
         can lead to unwanted behavior if you stream big data.
 
         This behavior can be disabled by setting
-        :attr:`implicit_iterable_buffering` to `False`.
+        :attr:`implicit_seqence_conversion` to `False`.
         """
-        self._ensure_buffered_iterable()
-        return ''.join(self.iterable)
+        self._ensure_sequence()
+        return ''.join(self.iter_encoded())
     def _set_data(self, value):
         # if an unicode string is set, it's encoded directly.  this allows
         # us to guess the content length automatically in `get_wsgi_headers`.
         if isinstance(value, unicode):
             value = value.encode(self.charset)
-        self.iterable = [value]
+        self.response = [value]
     data = property(_get_data, _set_data, doc=_get_data.__doc__)
     del _get_data, _set_data
 
-    def _ensure_buffered_iterable(self):
-        if self.iterable_is_buffered:
+    def _ensure_sequence(self, mutable=False):
+        """This method can be called by methods that need a sequence.  If
+        `mutable` is true, it will also ensure that the response sequence
+        is a standard Python list.
+
+        .. versionadded:: 0.6
+        """
+        if self.is_sequence:
+            # if we need a mutable object, we ensure it's a list.
+            if mutable and not isinstance(self.response, list):
+                self.response = list(self.response)
             return
-        if not self.implicit_iterable_buffering:
-            raise RuntimeError('The response object required the iterator '
-                               'to be buffered but implicit iterator '
-                               'consumption was disabled')
-        self.buffer_iterator()
+        if not self.implicit_seqence_conversion:
+            raise RuntimeError('The response object required the iterable '
+                               'to be a sequence, but the implicit '
+                               'conversion was disabled.  Call '
+                               'make_sequence() yourself.')
+        self.make_sequence()
 
-    @property
-    def iterable_is_buffered(self):
-        """If the iterator is buffered, this property will be `True`.  A
-        response object will consider an iterator to be buffered if the
-        response attribute is a list or tuple.
-        """
-        return isinstance(self.iterable, (tuple, list))
+    def make_sequence(self):
+        """Converts the response iterator in a list.  By default this happens
+        automatically if required.  If `implicit_seqence_conversion` is
+        disabled, this method is not automatically called and some properties
+        might raise exceptions.  This also encodes all the items.
 
-    def buffer_iterator(self):
-        """Buffers the response iterator in a list.  By default this happens
-        if required.  If `implicit_iterable_buffering` is disabled, this
-        method is not automatically called and some properties might raise
-        exceptions.  This also encodes all the items.
+        .. versionadded:: 0.6
         """
-        if not self.iterable_is_buffered:
+        if not self.is_sequence:
             # if we consume an iterable we have to ensure that the close
             # method of the iterable is called if available when we tear
             # down the response
-            close = getattr(self.iterable, 'close', None)
-            self.iterable = list(self.iter_encoded())
+            close = getattr(self.response, 'close', None)
+            self.response = list(self.iter_encoded())
             if close is not None:
                 self.call_on_close(close)
 
     def iter_encoded(self, charset=None):
-        """Iter the response encoded with the encoding specified.  If no
-        encoding is given the encoding from the class is used.  Note that
-        this does not encode data that is already a bytestring.  If the
-        response object is invoked as WSGI application the return value
-        of this method is used as application iterator except if
+        """Iter the response encoded with the encoding of the response.
+        If the response object is invoked as WSGI application the return
+        value of this method is used as application iterator unless
         :attr:`direct_passthrough` was activated.
+
+        .. versionchanged:: 0.6
+
+           The `charset` parameter was deprecated and became a no-op.
         """
-        charset = charset or self.charset
-        for item in self.iterable:
+        # XXX: deprecated
+        if __debug__ and charset is not None:
+            from warnings import warn
+            warn(DeprecationWarning('charset was deprecated and is ignored.'),
+                 stacklevel=2)
+        charset = self.charset
+        if __debug__:
+            _warn_if_string(self.response)
+        for item in self.response:
             if isinstance(item, unicode):
                 yield item.encode(charset)
             else:
@@ -780,31 +794,42 @@ class BaseResponse(object):
     @property
     def header_list(self):
         # XXX: deprecated
-        from warnings import warn
-        warn(DeprecationWarning('header_list is deprecated'),
-             stacklevel=2)
+        if __debug__:
+            from warnings import warn
+            warn(DeprecationWarning('header_list is deprecated'),
+                 stacklevel=2)
         return self.headers.to_list(self.charset)
 
     @property
     def is_streamed(self):
-        """If the response is streamed (the response is not a sequence) this
-        property is `True`.  In this case streamed means that there is no
-        information about the number of iterations.  This is usully `True`
-        if a generator is passed to the response object.
+        """If the response is streamed (the response is not a iterable with
+        a length information) this property is `True`.  In this case streamed
+        means that there is no information about the number of iterations.
+        This is usully `True` if a generator is passed to the response object.
 
         This is useful for checking before applying some sort of post
         filtering that should not take place for streamed responses.
         """
         try:
-            len(self.iterable)
+            len(self.response)
         except TypeError:
             return True
         return False
 
+    @property
+    def is_sequence(self):
+        """If the iterator is buffered, this property will be `True`.  A
+        response object will consider an iterator to be buffered if the
+        response attribute is a list or tuple.
+
+        .. versionadded:: 0.6
+        """
+        return isinstance(self.response, (tuple, list))
+
     def close(self):
         """Close the wrapped response if possible."""
-        if hasattr(self.iterable, 'close'):
-            self.iterable.close()
+        if hasattr(self.response, 'close'):
+            self.response.close()
         for func in self._on_close:
             func()
 
@@ -816,15 +841,18 @@ class BaseResponse(object):
         .. versionchanged:: 0.6
            The `Content-Length` header is now set.
         """
-        self.buffer_iterator()
-        self.headers['Content-Length'] = str(len(self.data))
+        # we explicitly set the length to a list of the *encoded* response
+        # iterator.  Even if the implicit sequence conversion is disabled.
+        self.response = list(self.iter_encoded())
+        self.headers['Content-Length'] = str(sum(map(len, self.response)))
 
     def fix_headers(self, environ):
         # XXX: deprecated
-        from warnings import warn
-        warn(DeprecationWarning('called into deprecated fix_headers baseclass '
-                                'method.  Use get_wsgi_headers instead.'),
-             stacklevel=2)
+        if __debug__:
+            from warnings import warn
+            warn(DeprecationWarning('called into deprecated fix_headers baseclass '
+                                    'method.  Use get_wsgi_headers instead.'),
+                 stacklevel=2)
         self.headers[:] = self.get_wsgi_headers(environ)
 
     def get_wsgi_headers(self, environ):
@@ -877,15 +905,15 @@ class BaseResponse(object):
         # should try to do that.  But only if this does not involve
         # flattening the iterator or encoding of unicode strings in
         # the response.
-        if self.iterable_is_buffered and 'content-length' not in self.headers:
+        if self.is_sequence and 'content-length' not in self.headers:
             try:
-                content_length = sum(len(str(x)) for x in self.iterable)
+                content_length = sum(len(str(x)) for x in self.response)
             except UnicodeError:
                 # aha, something non-bytestringy in there, too bad, we
                 # can't savely figure out the length of the response.
                 pass
             else:
-                self.headers['Content-Length'] = str(content_length)
+                headers['Content-Length'] = str(content_length)
 
         return headers
 
@@ -907,7 +935,9 @@ class BaseResponse(object):
            100 <= self.status_code < 200 or self.status_code in (204, 304):
             return ()
         if self.direct_passthrough:
-            return self.iterable
+            if __debug__:
+                _warn_if_string(self.response)
+            return self.response
         return self.iter_encoded()
 
     def get_wsgi_response(self, environ):
@@ -925,7 +955,7 @@ class BaseResponse(object):
         """
         # XXX: code for backwards compatibility with custom fix_headers
         # methods.
-        if self.fix_headers.func_code is not \
+        if __debug__ and self.fix_headers.func_code is not \
            BaseResponse.fix_headers.func_code:
             from warnings import warn
             warn(DeprecationWarning('fix_headers changed behavior in 0.6 '
@@ -1159,8 +1189,8 @@ class ResponseStream(object):
     def write(self, value):
         if self.closed:
             raise ValueError('I/O operation on closed file')
-        self.response._ensure_buffered_iterable()
-        self.response.iterable.append(value)
+        self.response._ensure_sequence(mutable=True)
+        self.response.response.append(value)
 
     def writelines(self, seq):
         for item in seq: